2513 update token limit config with schema generation (#2514)
* refactor default values in configuration fields to use String type; update related classes and add new token limit configuration files

* refactor default values in configuration fields to use String type; update related classes and add new token limit configuration files

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

+++ token-limit/src/test/java/com/networknt/token/limit/TokenLimitConfigTest.java
@@ -17,7 +17,7 @@ public class TokenLimitConfigTest {
     public void testConfigData() {
         Assert.assertTrue(config.isEnabled());
         Assert.assertTrue(config.isErrorOnLimit());
-        Assert.assertEquals(2, config.getDuplicateLimit());
+        Assert.assertEquals(2, config.getDuplicateLimit().intValue());
         Assert.assertEquals("expires_in", config.getExpireKey());
         Assert.assertTrue(config.getTokenPathTemplates().contains("/oauth2/(?<instanceId>[^/]+)/v1/token"));
     }
